complete Result service

+++ Result/exam/models.py
@@ -4,7 +4,7 @@ from topic.models import TopicModel
 from question.models import QuestionModel
 
 class ExamModel(models.Model):
-    examId = models.BigAutoField(primary_key=True)
+    examId = models.IntegerField(primary_key=True)
     name = models.CharField(max_length=50)
     numberForCorrectAnswer = models.IntegerField(default=0)
     numberForIncorrectAnswer = models.IntegerField(default=0)
